Refactor package nesting
