+++ kaldb/src/main/java/com/slack/kaldb/chunk/manager/RollOverChunkTask.java
@@ -1,4 +1,4 @@
-package com.slack.kaldb.chunk.manager.indexing;
+package com.slack.kaldb.chunk.manager;
 
 import com.slack.kaldb.blobfs.s3.S3BlobFs;
 import com.slack.kaldb.chunk.ReadWriteChunkImpl;
