+++ kaldb/src/main/java/com/slack/kaldb/server/Kaldb.java
@@ -3,9 +3,9 @@ package com.slack.kaldb.server;
 import com.google.common.util.concurrent.MoreExecutors;
 import com.google.common.util.concurrent.Service;
 import com.google.common.util.concurrent.ServiceManager;
-import com.slack.kaldb.chunk.manager.caching.CachingChunkManager;
-import com.slack.kaldb.chunk.manager.indexing.ChunkCleanerService;
-import com.slack.kaldb.chunk.manager.indexing.IndexingChunkManager;
+import com.slack.kaldb.chunk.manager.CachingChunkManager;
+import com.slack.kaldb.chunk.manager.ChunkCleanerService;
+import com.slack.kaldb.chunk.manager.IndexingChunkManager;
 import com.slack.kaldb.config.KaldbConfig;
 import com.slack.kaldb.logstore.LogMessage;
 import com.slack.kaldb.logstore.search.KaldbDistributedQueryService;
