+++ kaldb/src/test/java/com/slack/kaldb/chunk/IndexingChunkManagerTest.java
@@ -1,10 +1,10 @@
 package com.slack.kaldb.chunk;
 
-import static com.slack.kaldb.chunk.manager.indexing.IndexingChunkManager.LIVE_BYTES_INDEXED;
-import static com.slack.kaldb.chunk.manager.indexing.IndexingChunkManager.LIVE_MESSAGES_INDEXED;
-import static com.slack.kaldb.chunk.manager.indexing.RollOverChunkTask.ROLLOVERS_COMPLETED;
-import static com.slack.kaldb.chunk.manager.indexing.RollOverChunkTask.ROLLOVERS_FAILED;
-import static com.slack.kaldb.chunk.manager.indexing.RollOverChunkTask.ROLLOVERS_INITIATED;
+import static com.slack.kaldb.chunk.manager.IndexingChunkManager.LIVE_BYTES_INDEXED;
+import static com.slack.kaldb.chunk.manager.IndexingChunkManager.LIVE_MESSAGES_INDEXED;
+import static com.slack.kaldb.chunk.manager.RollOverChunkTask.ROLLOVERS_COMPLETED;
+import static com.slack.kaldb.chunk.manager.RollOverChunkTask.ROLLOVERS_FAILED;
+import static com.slack.kaldb.chunk.manager.RollOverChunkTask.ROLLOVERS_INITIATED;
 import static com.slack.kaldb.config.KaldbConfig.DEFAULT_START_STOP_DURATION;
 import static com.slack.kaldb.logstore.LuceneIndexStoreImpl.MESSAGES_FAILED_COUNTER;
 import static com.slack.kaldb.logstore.LuceneIndexStoreImpl.MESSAGES_RECEIVED_COUNTER;
@@ -22,10 +22,10 @@ import com.google.common.util.concurrent.ListeningExecutorService;
 import com.google.common.util.concurrent.MoreExecutors;
 import com.slack.kaldb.blobfs.s3.S3BlobFs;
 import com.slack.kaldb.chunk.manager.ChunkManager;
-import com.slack.kaldb.chunk.manager.indexing.ChunkCleanerService;
-import com.slack.kaldb.chunk.manager.indexing.ChunkRollOverStrategy;
-import com.slack.kaldb.chunk.manager.indexing.ChunkRollOverStrategyImpl;
-import com.slack.kaldb.chunk.manager.indexing.IndexingChunkManager;
+import com.slack.kaldb.chunk.manager.ChunkCleanerService;
+import com.slack.kaldb.chunk.manager.ChunkRollOverStrategy;
+import com.slack.kaldb.chunk.manager.ChunkRollOverStrategyImpl;
+import com.slack.kaldb.chunk.manager.IndexingChunkManager;
 import com.slack.kaldb.com.slack.kaldb.logstore.search.AlreadyClosedLogIndexSearcherImpl;
 import com.slack.kaldb.com.slack.kaldb.logstore.search.IllegalArgumentLogIndexSearcherImpl;
 import com.slack.kaldb.logstore.LogMessage;
