+++ kaldb/src/test/java/com/slack/kaldb/logstore/search/KaldbLocalQueryServiceTest.java
@@ -9,8 +9,8 @@ import static org.assertj.core.api.Assertions.catchThrowable;
 import brave.Tracing;
 import com.adobe.testing.s3mock.junit4.S3MockRule;
 import com.google.protobuf.ByteString;
-import com.slack.kaldb.chunk.manager.indexing.IndexingChunkManager;
-import com.slack.kaldb.chunk.manager.indexing.RollOverChunkTask;
+import com.slack.kaldb.chunk.manager.IndexingChunkManager;
+import com.slack.kaldb.chunk.manager.RollOverChunkTask;
 import com.slack.kaldb.logstore.LogMessage;
 import com.slack.kaldb.logstore.LogWireMessage;
 import com.slack.kaldb.proto.service.KaldbSearch;
