+++ kaldb/src/test/java/com/slack/kaldb/server/KaldbIndexerTest.java
@@ -17,8 +17,8 @@ import com.linecorp.armeria.server.Server;
 import com.linecorp.armeria.server.ServerBuilder;
 import com.linecorp.armeria.server.grpc.GrpcService;
 import com.linecorp.armeria.server.grpc.GrpcServiceBuilder;
-import com.slack.kaldb.chunk.manager.indexing.IndexingChunkManager;
-import com.slack.kaldb.chunk.manager.indexing.RollOverChunkTask;
+import com.slack.kaldb.chunk.manager.IndexingChunkManager;
+import com.slack.kaldb.chunk.manager.RollOverChunkTask;
 import com.slack.kaldb.config.KaldbConfig;
 import com.slack.kaldb.logstore.LogMessage;
 import com.slack.kaldb.logstore.search.KaldbLocalQueryService;
