+++ kaldb/src/test/java/com/slack/kaldb/testlib/ChunkManagerUtil.java
@@ -7,9 +7,9 @@ import com.google.common.io.Files;
 import com.google.common.util.concurrent.MoreExecutors;
 import com.slack.kaldb.blobfs.s3.S3BlobFs;
 import com.slack.kaldb.chunk.SearchContext;
-import com.slack.kaldb.chunk.manager.indexing.ChunkRollOverStrategy;
-import com.slack.kaldb.chunk.manager.indexing.ChunkRollOverStrategyImpl;
-import com.slack.kaldb.chunk.manager.indexing.IndexingChunkManager;
+import com.slack.kaldb.chunk.manager.ChunkRollOverStrategy;
+import com.slack.kaldb.chunk.manager.ChunkRollOverStrategyImpl;
+import com.slack.kaldb.chunk.manager.IndexingChunkManager;
 import com.slack.kaldb.proto.config.KaldbConfigs;
 import com.slack.kaldb.server.MetadataStoreService;
 import io.micrometer.core.instrument.MeterRegistry;
